Put helper functions into helpers.js
#227
